+++ src/market-bot.ts
@@ -1,4 +1,4 @@
-import { setContext, captureException } from '@sentry/node';
+import Bugsnag from '@bugsnag/js';
 import elastic from 'elastic-apm-node';
 import { createConnection } from 'typeorm';
 
@@ -195,7 +195,7 @@ async function processMessage(message: Message, transaction: any): Promise<void>
 }
 
 export function handleError(message: Message, caughtError: Error) {
-    setContext('command', {command: message.content});
-    captureException(caughtError);
+    Bugsnag.addMetadata('command', {command: message.content});
+    Bugsnag.notify(caughtError);
     message.sendError(caughtError).then();
 }